Les prix sont en float pas en int

+++ src/Imerir/NoyauBundle/Controller/SoapController.php
@@ -246,6 +246,7 @@ class SoapController extends ContainerAware
 	 *   ----> ...
 	 * 
 	 * @param articles Un chaine de caractère JSON correspondant à un tableau ayant le format décrit ci-dessus.
+	 * @param avecPrix Dit qu'on fournit le prix de l'article ou non pendant l'inventaire
 	 * 
 	 * @Soap\Method("faireInventaire")
 	 * @Soap\Param("articles",phpType="string")
@@ -301,9 +302,9 @@ class SoapController extends ContainerAware
 				$resultat = $pdo->query($sql);
 			}
 			
-			if ($avecPrix) {
+			if ($avecPrix) { // Si on enregistre le prix avec
 				$sql = 'INSERT INTO prix(ref_article, montant_fournisseur, montant_client, date_modif)
-							VALUE (\''.$idArticle.'\', 0, \''.(int)$prixClient.'\', NOW())';
+							VALUE (\''.$idArticle.'\', 0, \''.(float)$prixClient.'\', NOW())';
 				$resultat = $pdo->query($sql);
 			}
